Python: renamed documentation generator, and making use of string output from configuration writer
